Added vectors to portfolio
I have added three vectors, one with the ticker names, one with the
variances, and one with the mean returns. Maybe we should have these in
a csv file.
Actually, runnung the choose(500, 2) on these vectors takes about 60
seconds, so maybe a csv of the final calculation would be better, this
is for anybody in the group to use without waiting 10 minutes to get
these numbers